requirements.txt
numpy==1.24.2                     # Versión específica de numpy
opencv-contrib-python==4.10.0.84  # Versión de OpenCV con módulos extra (incluye ArUco)
opencv-python==4.10.0.84          # Versión principal de OpenCV
scipy==1.10.1                     # SciPy para cálculos científicos
matplotlib==3.7.1                 # Matplotlib para visualizaciones
PyQt5==5.15.9                     # PyQt5 para la interfaz gráfica
PyOpenGL==3.1.6                   # PyOpenGL para la renderización de los objetos 3D
pygame==2.6.1                     # Pygame para la interfaz gráfica